requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
pymongo>=4.3.0
plotly>=5.14.0
python-dotenv>=1.0.0
yfinance>=0.2.18
numpy>=1.23.0
matplotlib>=3.7.0
scipy>=1.10.0